+++ src/cxx/VulkanLib/Device/Descriptors/DescriptorPool.cpp
@@ -25,7 +25,7 @@ DescriptorPool::DescriptorPool(std::shared_ptr<LogicalDevice> logicalDevice, boo
     vk::DescriptorPoolCreateInfo poolInfo{};
     poolInfo.maxSets = 1500;
     poolInfo.pPoolSizes = poolSizes.data();
-    poolInfo.poolSizeCount = poolSizes.size();
+    poolInfo.poolSizeCount =(uint32_t) poolSizes.size();
     descriptorPool = logicalDevice->getDevice().createDescriptorPool(poolInfo);
 }
